provisioning/roles/deploy/files/app/api/requirements.txt
flask==2.3.0
flask-cors==4.0.0
psycopg2-binary==2.9.6
